Apply KCEJ source changes
Replace memory allocation with a bump allocator
Remove zalloc from z_stream
Remove zfree from z_stream
Move data_type and adler before opaque in z_stream
Redefine uLong as unsigned int
Use static data instead of allocating codes struct
Use stack arrays instead of allocation for huft work area
Remove BUILDFIXED path for inflate_trees_fixed

+++ module/zlibdec/zconf.h
@@ -214,7 +214,7 @@
 typedef unsigned char  Byte;  /* 8 bits */
 #endif
 typedef unsigned int   uInt;  /* 16 bits or more */
-typedef unsigned long  uLong; /* 32 bits or more */
+typedef unsigned int   uLong; /* 32 bits or more */
 
 #ifdef SMALL_MEDIUM
    /* Borland C/C++ and some old MSC versions ignore FAR inside typedef */
